refactor(transit): move filterBoardable to timetable-filter module
Consolidate filter functions in `timetable-filter.ts`. `filterBoardable`
was defined in `timetable-utils.ts` next to entry-level predicates but
its shape (TimetableEntry[] → TimetableEntry[]) belongs with the other
filter wrappers (`prepareStopTimetable`, `prepareRouteHeadsignTimetable`,
`filterByAgency`, `filterByRouteType`).

`hasBoardable(entries)` stays in `timetable-utils.ts` because it returns
a boolean — same shape as the other `is*` / `has*` predicates.
`isDropOffOnly` (used internally by `filterBoardable`) also stays as the
canonical "non-boardable" entry-level predicate.

Also relocate the `filterBoardable` test describe from
`timetable-utils.test.ts` to `timetable-filter.test.ts`. As a side
effect of removing the now-redundant `describe('filterBoardable')`
wrapper from the utils test file, four describes that had been
accidentally nested inside it (`getDisplayMinutes`, `getStopServiceState`,
`getTimetableEntriesState`, `getFilteredTimetableEntriesState`) become
correctly top-level.

Pure refactor: no behavior change. 2929 tests pass, typecheck and
prettier clean.
